pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "splendor"
version = "0.1.0a0"
description = "Local-first, git-native, schema-driven knowledge compiler for code-and-research repositories."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [
  { name = "Splendor contributors" },
]
dependencies = [
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0.2,<7.0.0",
]

[dependency-groups]
dev = [
  "coverage[toml]>=7.6,<8.0",
  "pre-commit>=4.0,<5.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
  "ruff>=0.11.0,<0.12.0",
]

[project.scripts]
splendor = "splendor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/splendor"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "B",
  "E",
  "F",
  "I",
  "UP",
]

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
relative_files = true
source = ["splendor"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.coverage.paths]
source = [
  "src",
  "/Users/runner/work/splendor/splendor/src",
  "/home/runner/work/splendor/splendor/src",
]
